Adds some dropdownHelper tests and moves checkIfItemIsSelected.js to that file
